chore: moved chatPage to 'apps/client/src/pages'
